graph was not plotting properly after test had completed. fixed
