+++ src/lib/histogramCache.json
@@ -1,1 +1,1 @@
-[]
+[{"x":[0,10,20,30,40,50,55.00000000000001,60,65,70,75,77.5,80,82.5,85,87.5,88.75,90,91.25,92.5,93.75,94.375,95,95.625,96.25,96.875,97.1875,97.5,97.8125,98.125,98.4375,98.5938,98.75,98.9062,99.0625,99.2188,99.2969,99.375,99.4531,99.5313,99.60940000000001,99.64840000000001,99.6875,99.7266,99.76559999999999,99.8047],"y":[0.507,0.985,1.18,1.392,1.579,1.805,1.944,2.046,2.183,2.269,2.409,2.507,2.669,2.861,3.029,3.357,3.559,3.747,3.941,4.283,4.487,4.651,4.935,5.547,5.803,6.235,6.259,6.575,6.867,7.451,7.607,7.607,7.667,8.743,8.743,9.951,9.951,9.951,10.215,10.215,12.055,12.055,12.055,12.055,12.055,12.527],"type":"scatter","mode":"lines+markers","hovertemplate":"X: %{x}%<br>Y: %{y}ms","marker":{"color":"rgb(203, 78, 219)"}}]
